pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seizure-forecasting"
version = "1.0.0"
description = "Seizure forecasting from scalp EEG using deep learning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9,<3.12"
authors = [
    {name = "Seizure Forecasting Team"}
]
keywords = ["eeg", "seizure", "prediction", "forecasting", "deep-learning", "medical"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    # Core scientific stack
    "numpy>=1.24.0,<2.0.0",
    "scipy>=1.10.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    
    # EEG processing
    "mne>=1.5.0",
    "pyedflib>=0.1.30",
    
    # Deep learning
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    
    # ML extras
    "xgboost>=2.0.0",
    
    # Configuration and logging
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    
    # Demo app
    "streamlit>=1.28.0",
    
    # Data storage
    "h5py>=3.9.0",
    
    # Utilities
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
verify-dataset = "scripts.verify_dataset:main"
build-cache = "scripts.build_cache:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*", "app*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
